% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctapply.R
\name{ctapply}
\alias{ctapply}
\title{Standard Central Tendency function}
\usage{
ctapply(x, np = TRUE, print = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{np}{logical TRUE/FALSE:  The user has a choice to which measure to calculate}

\item{print}{logical TRUE/FALSE: The user has a choice to print the result}
}
\description{
This function can be used to get the measure of central tendency and spread for a numeric vector x
}
\examples{
ctapply()

#Sample data used detail: Cost Function of Electricity Producers (1955)
#The sample data was taken from: https://vincentarelbundock.github.io/Rdatasets/datasets.html
#Electricity1955 <- read.table("C:\\USF\\SPRING2022\\R-PROGRAMMING\\Assignment11\\Electricity1955.csv", header = TRUE, sep=",")
#electricdata <- data.frame(Electricity1955)
#electricdata
#save(electricdata, file = "data\\electricdata.RData")


# Function description: get the measure of central tendency
}
\keyword{ctapply}
