% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctbarplot.R
\name{ctbarplot}
\alias{ctbarplot}
\title{Plot bar graph}
\usage{
ctbarplot(x)
}
\arguments{
\item{x=cost}{numeric vector (could use package data electricdata$cost)}
}
\description{
You can either use the electric data that comes with the package or generate your own
}
\examples{
ctbarplot()
}
\keyword{ctbarplot}
