% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctpairplot.R
\name{ctpairplot}
\alias{ctpairplot}
\title{Plot scattere pair graph}
\usage{
ctpairplot(x)
}
\arguments{
\item{x=}{electricdata data frame}
}
\description{
You can either use the electric data that comes with the package or generate your own
}
\examples{
ctpairplot()
}
\keyword{ctpairplot}
