% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctscattergraph.R
\name{ctscattergraph}
\alias{ctscattergraph}
\title{Plot histogram}
\usage{
ctscattergraph(x, y)
}
\arguments{
\item{x}{= cost numeric vector (could use package data electricdata$cost)}

\item{y}{= fuel numeric vector (could use package data electricdata$fuel)}
}
\description{
You may use the electric data that comes with the package
}
\examples{
ctscattergraph()
}
\keyword{ctscattergraph}
